feat : Space 모듈 Service 추가 - 공간생성 - 공간수정 - 공간삭제 - 공간상세내역 - 공간전체조회

+++ src/main/java/com/example/room/space/service/SpaceService.java
@@ -1,11 +1,11 @@
 package com.example.room.space.service;
 
+import com.example.room.reservation.service.ReservationService;
 import com.example.room.space.dto.RequestSpace;
 import com.example.room.space.dto.ResponseSpace;
 import com.example.room.space.entity.space.Space;
 import com.example.room.space.repository.SpaceRepository;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -19,40 +19,43 @@ public class SpaceService {
 
     private final SpaceRepository spaceRepository;
 
+    private final ReservationService reservationService;
+
     @Transactional
     public ResponseSpace create(RequestSpace requestSpace) {
-        Space space = spaceRepository.save(requestSpace.toEntity());
-        space.updateRental(requestSpace.getRentalDto());
-        space.updateFacility(requestSpace.getFacilityDto());
-        return space.toDto();
+        Space space = requestSpace.toEntity();
+        return new ResponseSpace(spaceRepository.save(space));
     }
 
     @Transactional(readOnly = true)
-    public ResponseSpace findBySpaceId(long id) {
-        return spaceRepository.findById(id).orElseThrow(() -> {
+    public ResponseSpace findBySpaceId(Long id) {
+        return new ResponseSpace(spaceRepository.findById(id).orElseThrow(() -> {
             throw new IllegalArgumentException("조회할 수 있는 공간이 없습니다.");
-        }).toDto();
+        }));
     }
 
     @Transactional(readOnly = true)
     public List<ResponseSpace> findAll(Pageable pageable) {
         Page<Space> spaces = spaceRepository.findAll(pageable);
         return spaces.stream()
-            .map(Space::toDto)
+            .map(ResponseSpace::new)
             .collect(Collectors.toList());
     }
 
     @Transactional
-    public void deleteBySpaceId(long id) {
+    public void deleteBySpaceId(Long id) {
+        //TODO
+        if (reservationService.isHoldReservationBySpaceId(id)) {
+            throw new IllegalArgumentException("예약이 예정된 공간입니다. 예약이 있는 경우 삭제하실 수 없습니다.");
+        }
         spaceRepository.deleteById(id);
     }
 
     @Transactional
     public ResponseSpace update(RequestSpace requestSpace) {
-        Space space = spaceRepository.findSpaceById(requestSpace.getId()).orElseThrow(() -> {
-            throw new IllegalArgumentException("수정할 수 있는 공간이 없습니다.");
-        });
-        space.update(requestSpace);
-        return space.toDto();
+        return new ResponseSpace(
+            spaceRepository.findSpaceById(requestSpace.getId()).orElseThrow(() -> {
+                throw new IllegalArgumentException("수정할 수 있는 공간이 없습니다.");
+            }));
     }
 }
